requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.54.0
anyio==4.9.0
backoff==2.2.1
beautifulsoup4==4.13.4
boto3==1.38.34
botocore==1.38.34
browser-use==0.2.6
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
distro==1.9.0
faiss-cpu==1.11.0
filetype==1.2.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-aws==0.2.25
langchain-core==0.3.64
langchain-deepseek==0.1.3
langchain-google-genai==2.1.5
langchain-ollama==0.3.3
langchain-openai==0.3.21
langchain-text-splitters==0.3.8
langsmith==0.3.45
markdownify==1.1.0
mem0ai==0.1.107
numpy==2.3.0
ollama==0.5.1
openai==1.86.0
orjson==3.10.18
packaging==24.2
patchright==1.52.5
playwright==1.52.0
portalocker==2.10.1
posthog==4.7.0
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pyee==13.0.0
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
qdrant-client==1.14.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
s3transfer==0.13.0
screeninfo==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uuid7==0.1.0
zstandard==0.23.0
